Changed: Apply Template pattern in `StringGenerator`
The classes extending `StringGenerator` has same form each other in `getMarkdownString()`, so applied Template pattern for standardizing MarkdownString.
