+++ termux-shared/src/main/java/com/termux/shared/net/socket/local/StringGenerator/PeerCredStringGenerator.java
@@ -23,23 +23,17 @@ public class PeerCredStringGenerator extends StringGenerator{
     }
 
     @Override
-    public String getMarkdownString() {
-        StringBuilder markdownString = new StringBuilder();
-
-        markdownString.append("## ").append("Peer Cred");
-
-        for (Pair<String, Object> logVar: getLogVariableList()) {
-            String label = logVar.first;
-            Object object = logVar.second;
-            switch(label) {
-                case "Cmdline":
-                    if (label != null) markdownString.append("\n").append(MarkdownUtils.getMultiLineMarkdownStringEntry(label, object, "-"));
-                    break;
-                default:
-                    markdownString.append("\n").append(MarkdownUtils.getSingleLineMarkdownStringEntry(label, object, "-"));
-            }
-        }
-
-        return markdownString.toString();
+    public void appendMarkdownHeader(StringBuilder markdownString) {
+        markdownString.append("Peer Cred:");
     }
+
+    @Override
+    public boolean isMarkdownMultiLine(String label) {
+        return label.equals("Cmdline");
+    }
+
+    @Override
+    public void appendMarkdownFooter(StringBuilder markdownString) {
+    }
+
 }
