+++ termux-shared/src/main/java/com/termux/shared/net/socket/local/StringGenerator/StringGenerator.java
@@ -3,6 +3,7 @@ package com.termux.shared.net.socket.local.StringGenerator;
 import android.util.Pair;
 
 import com.termux.shared.logger.Logger;
+import com.termux.shared.markdown.MarkdownUtils;
 
 import java.util.List;
 
@@ -41,7 +42,27 @@ public abstract class StringGenerator {
     public abstract boolean isLogMultiLine(String label);
     public abstract void appendLogFooter(StringBuilder logString);
 
-    public abstract String getMarkdownString();
+    public String getMarkdownString() {
+        StringBuilder markdownString = new StringBuilder();
 
+        appendMarkdownHeader(markdownString);
+
+        for (Pair<String, Object> logVar: getLogVariableList()) {
+            String label = logVar.first;
+            Object object = logVar.second;
+            if (isMarkdownMultiLine(label))
+                markdownString.append("\n").append(MarkdownUtils.getMultiLineMarkdownStringEntry(label, object, "-"));
+            else
+                markdownString.append("\n").append(MarkdownUtils.getSingleLineMarkdownStringEntry(label, object, "-"));
+        }
+
+        appendMarkdownFooter(markdownString);
+
+        return markdownString.toString();
+    }
+
+    public abstract void appendMarkdownHeader(StringBuilder logString);
+    public abstract boolean isMarkdownMultiLine(String label);
+    public abstract void appendMarkdownFooter(StringBuilder logString);
 
 }
